Step 5: Define Attributes and add Inspector controls
Here we define attributes within `registerBlockType`. The attributes are the dynamic values that will be used to modify how the block behaves.

We also add our Inspector controls, which are form fields that can be used to interact with the block. Right now, we can interact with the controls, but they don't actually save any data. We'll work on that in the next step.